Implement initial JS and small html and css update
The JavaScript code added will allow a winner to be decided based on
what the user picks and what the computer randomly chooses. A win-count
for the player and computer is also added and whoever wins 5 rounds
first wins the match.

JavaScript code still needs to be added that actually ends the game
and resets as if the webpage was just loaded.

 Changes to be committed:
	modified:   css/style.css
	modified:   index.html
	modified:   index.js